Abstract state machine testing
