+++ hedgehog-example/hedgehog-example.cabal
@@ -45,6 +45,7 @@ test-suite test
     , base                            >= 3          && < 5
     , containers                      >= 0.4        && < 0.6
     , filepath                        >= 1.3        && < 1.4
+    , hashtables                      >= 1.2        && < 1.3
     , mmorph                          >= 1.0        && < 1.1
     , mtl                             >= 2.1        && < 2.3
     , parsec                          >= 3.1        && < 3.2
@@ -53,4 +54,4 @@ test-suite test
     , resourcet                       >= 1.1        && < 1.2
     , temporary-resourcet             >= 0.1        && < 0.2
     , text                            >= 1.1        && < 1.3
-    , transformers                    >= 0.3        && < 0.6
+    , transformers                    >= 0.5        && < 0.6
